+++ src/assets.rs
@@ -182,6 +182,16 @@ impl HighlightingAssets {
                 Err(Error::UndetectedSyntax(path.to_string_lossy().into()))
             }
 
+            Some(MappingTarget::MapToUnknownUnlessExactFileNameMatch) => {
+                let file_name = path.file_name().unwrap_or_default();
+                // Check if we have an exact match using file name.
+                Ok(self.find_syntax_by_name(file_name.to_str().unwrap())?
+                    // if we do not check to see if we have an exact match using extension.
+                    .ok_or_else(|| self.get_extension_syntax(file_name)?
+                        // If no matches fail like MapToUnknown
+                        .ok_or_else(|| Error::UndetectedSyntax(path.to_string_lossy().into()))).unwrap())
+            }
+
             Some(MappingTarget::MapTo(syntax_name)) => self
                 .find_syntax_by_name(syntax_name)?
                 .ok_or_else(|| Error::UnknownSyntax(syntax_name.to_owned())),