+++ src/syntax_mapping.rs
@@ -8,6 +8,7 @@ use globset::{Candidate, GlobBuilder, GlobMatcher};
 pub enum MappingTarget<'a> {
     MapTo(&'a str),
     MapToUnknown,
+    MapToUnknownUnlessExactFileNameMatch,
 }
 
 #[derive(Debug, Clone, Default)]
@@ -53,13 +54,9 @@ impl<'a> SyntaxMapping<'a> {
 
         // Nginx and Apache syntax files both want to style all ".conf" files
         // see #1131 and #1137
+	// Support general syntax highlighting for .conf files if exact match is found, resolves issue #1703
         mapping
-            .insert("*.conf", MappingTarget::MapToUnknown)
-            .unwrap();
-
-        // Re-insert a mapping for resolv.conf, see #1510
-        mapping
-            .insert("resolv.conf", MappingTarget::MapTo("resolv"))
+            .insert("*.conf", MappingTarget::MapToUnknownUnlessExactFileNameMatch)
             .unwrap();
 
         for glob in &[
@@ -79,16 +76,6 @@ impl<'a> SyntaxMapping<'a> {
             mapping
                 .insert(glob, MappingTarget::MapTo("Apache Conf"))
                 .unwrap();
-        }
-
-        // Support syntax highlighting for Tmux, resolves issue #1703
-        for glob in &[
-            "tmux.conf",
-	    "tmux",
-	    ".tmux.conf",
-	    ".tmux",
-        ] {
-            mapping.insert(glob, MappingTarget::MapTo("Tmux")).unwrap();
         }
 
         for glob in &[